Revert "Show notification when deactivate command is run in terminal (#22133)"
This reverts commit 66cea2169ef4c46a96e42725c4793ccce3b5c5df.
