+++ src/client/common/application/applicationShell.ts
@@ -10,7 +10,6 @@ import {
     DocumentSelector,
     env,
     Event,
-    EventEmitter,
     InputBox,
     InputBoxOptions,
     languages,
@@ -38,8 +37,7 @@ import {
     WorkspaceFolder,
     WorkspaceFolderPickOptions,
 } from 'vscode';
-import { traceError } from '../../logging';
-import { IApplicationShell, TerminalDataWriteEvent } from './types';
+import { IApplicationShell } from './types';
 
 @injectable()
 export class ApplicationShell implements IApplicationShell {
@@ -173,13 +171,5 @@ export class ApplicationShell implements IApplicationShell {
     }
     public createLanguageStatusItem(id: string, selector: DocumentSelector): LanguageStatusItem {
         return languages.createLanguageStatusItem(id, selector);
-    }
-    public get onDidWriteTerminalData(): Event<TerminalDataWriteEvent> {
-        try {
-            return window.onDidWriteTerminalData;
-        } catch (ex) {
-            traceError('Failed to get proposed API onDidWriteTerminalData', ex);
-            return new EventEmitter<TerminalDataWriteEvent>().event;
-        }
     }
 }
