+++ src/client/common/application/types.ts
@@ -67,17 +67,6 @@ import { Resource } from '../types';
 import { ICommandNameArgumentTypeMapping } from './commands';
 import { ExtensionContextKey } from './contextKeys';
 
-export interface TerminalDataWriteEvent {
-    /**
-     * The {@link Terminal} for which the data was written.
-     */
-    readonly terminal: Terminal;
-    /**
-     * The data being written.
-     */
-    readonly data: string;
-}
-
 export const IApplicationShell = Symbol('IApplicationShell');
 export interface IApplicationShell {
     /**
@@ -85,13 +74,6 @@ export interface IApplicationShell {
      * changes. The value of the event represents whether the window is focused.
      */
     readonly onDidChangeWindowState: Event<WindowState>;
-
-    /**
-     * An event which fires when the terminal's child pseudo-device is written to (the shell).
-     * In other words, this provides access to the raw data stream from the process running
-     * within the terminal, including VT sequences.
-     */
-    readonly onDidWriteTerminalData: Event<TerminalDataWriteEvent>;
 
     showInformationMessage(message: string, ...items: string[]): Thenable<string | undefined>;
