+++ src/client/terminals/serviceRegistry.ts
@@ -1,26 +1,25 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { IServiceManager } from '../ioc/types';
+import { interfaces } from 'inversify';
+import { ClassType } from '../ioc/types';
 import { TerminalAutoActivation } from './activation';
 import { CodeExecutionManager } from './codeExecution/codeExecutionManager';
 import { DjangoShellCodeExecutionProvider } from './codeExecution/djangoShellCodeExecution';
 import { CodeExecutionHelper } from './codeExecution/helper';
 import { ReplProvider } from './codeExecution/repl';
 import { TerminalCodeExecutionProvider } from './codeExecution/terminalCodeExecution';
-import {
-    ICodeExecutionHelper,
-    ICodeExecutionManager,
-    ICodeExecutionService,
-    ITerminalAutoActivation,
-    ITerminalEnvVarCollectionService,
-} from './types';
-import { TerminalEnvVarCollectionService } from './envCollectionActivation/service';
-import { IExtensionActivationService, IExtensionSingleActivationService } from '../activation/types';
-import { TerminalDeactivateLimitationPrompt } from './envCollectionActivation/deactivatePrompt';
-import { TerminalIndicatorPrompt } from './envCollectionActivation/indicatorPrompt';
+import { ICodeExecutionHelper, ICodeExecutionManager, ICodeExecutionService, ITerminalAutoActivation } from './types';
 
-export function registerTypes(serviceManager: IServiceManager): void {
+interface IServiceRegistry {
+    addSingleton<T>(
+        serviceIdentifier: interfaces.ServiceIdentifier<T>,
+        constructor: ClassType<T>,
+        name?: string | number | symbol,
+    ): void;
+}
+
+export function registerTypes(serviceManager: IServiceRegistry): void {
     serviceManager.addSingleton<ICodeExecutionHelper>(ICodeExecutionHelper, CodeExecutionHelper);
 
     serviceManager.addSingleton<ICodeExecutionManager>(ICodeExecutionManager, CodeExecutionManager);
@@ -38,17 +37,4 @@ export function registerTypes(serviceManager: IServiceManager): void {
     serviceManager.addSingleton<ICodeExecutionService>(ICodeExecutionService, ReplProvider, 'repl');
 
     serviceManager.addSingleton<ITerminalAutoActivation>(ITerminalAutoActivation, TerminalAutoActivation);
-    serviceManager.addSingleton<ITerminalEnvVarCollectionService>(
-        ITerminalEnvVarCollectionService,
-        TerminalEnvVarCollectionService,
-    );
-    serviceManager.addSingleton<IExtensionSingleActivationService>(
-        IExtensionSingleActivationService,
-        TerminalIndicatorPrompt,
-    );
-    serviceManager.addSingleton<IExtensionSingleActivationService>(
-        IExtensionSingleActivationService,
-        TerminalDeactivateLimitationPrompt,
-    );
-    serviceManager.addBinding(ITerminalEnvVarCollectionService, IExtensionActivationService);
 }
