+++ src/client/terminals/types.ts
@@ -33,11 +33,3 @@ export interface ITerminalAutoActivation extends IDisposable {
     register(): void;
     disableAutoActivation(terminal: Terminal): void;
 }
-
-export const ITerminalEnvVarCollectionService = Symbol('ITerminalEnvVarCollectionService');
-export interface ITerminalEnvVarCollectionService {
-    /**
-     * Returns true if we know with high certainity the terminal prompt is set correctly for a particular resource.
-     */
-    isTerminalPromptSetCorrectly(resource?: Resource): boolean;
-}
