--- typings/vscode-proposed/vscode.proposed.terminalDataWriteEvent.d.ts
@@ -1,31 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-declare module 'vscode' {
-    // https://github.com/microsoft/vscode/issues/78502
-    //
-    // This API is still proposed but we don't intent on promoting it to stable due to problems
-    // around performance. See #145234 for a more likely API to get stabilized.
-
-    export interface TerminalDataWriteEvent {
-        /**
-         * The {@link Terminal} for which the data was written.
-         */
-        readonly terminal: Terminal;
-        /**
-         * The data being written.
-         */
-        readonly data: string;
-    }
-
-    namespace window {
-        /**
-         * An event which fires when the terminal's child pseudo-device is written to (the shell).
-         * In other words, this provides access to the raw data stream from the process running
-         * within the terminal, including VT sequences.
-         */
-        export const onDidWriteTerminalData: Event<TerminalDataWriteEvent>;
-    }
-}
